Badge Cleanup and Documentation improvement
