+++ srv/app/flp/index.html
@@ -7,9 +7,9 @@
 	<meta http-equiv="Content-Type" content="text/html;charset=UTF-8" />
 	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 	<title>SAP Community Tools - Unsupported and for Fun</title>
-	<script src="https://sapui5.hana.ondemand.com/1.120.5/test-resources/sap/ushell/bootstrap/sandbox.js"></script>
+	<script src="https://sapui5.hana.ondemand.com/1.124.1/test-resources/sap/ushell/bootstrap/sandbox.js"></script>
 
-	<script src="https://sapui5.hana.ondemand.com/1.120.5/resources/sap-ui-core.js"
+	<script src="https://sapui5.hana.ondemand.com/1.124.1/resources/sap-ui-core.js"
 		data-sap-ui-libs="sap.m, sap.ushell, sap.collaboration, sap.ui.layout, sap.ui.rta, sap.uxap"
 		data-sap-ui-compatVersion="edge" data-sap-ui-theme="sap_horizon"
 		data-sap-ui-async="true"
