+++ srv/html/devtoberfest_header.html
@@ -5,7 +5,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width">
 
-    <title>Devtoberfest 2023</title>
+    <title>Devtoberfest 2024</title>
